Clarify CLI policy path ownership
